Extract core HMR client logic to shared runtime
This extracts the core HMR client logic, abstracting away hooks for platform-specific features like the websocket transport and page reload.

Changes:
- Create shared/runtime/hmr-runtime.ts (812 lines):
  - Extracted HMR implementation from browser dev-base.ts
  - Platform-agnostic HMR state management
  - Complete module.hot API implementation
  - HMR update flow logic with dependency tracking
- Refactor browser/runtime/base/dev-base.ts:
  - Remove HMR implementation (now in shared)
  - Add platform-specific hooks (websocket, page reload)
  - Use shared instantiateModuleShared function
  - Simplified to ~46 lines from ~685 lines
- Update shared/runtime/runtime-utils.ts:
  - Add SourceType enum and SourceData type
  - Add shared HMR state declarations
  - Add instantiateModuleShared function
- Update browser/runtime/base/runtime-base.ts:
  - Remove SourceType/SourceData (now in runtime-utils)
- Update browser HMR client to use shared implementation
- Remove unused browser/dev/hmr-client/index.ts
- Update Rust build to include hmr-runtime.ts in dev bundles
- Rerecord snapshots for HMR runtime changes

This prepares for implementing Node.js server HMR by making the HMR logic reusable across runtimes.

Test Plan: CI, existing HMR tests
