--- turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/dev/hmr-client/index.ts
@@ -1,1 +1,0 @@
-export * from './hmr-client'
