+++ turbopack/crates/turbopack-ecmascript-runtime/js/src/browser/runtime/base/dev-base.ts
@@ -14,11 +14,12 @@ const devContextPrototype = Context.prototype as TurbopackDevContext
  * It will be appended to the runtime code of each runtime right after the
  * shared runtime utils.
  */
-
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-const devModuleCache: ModuleCache<HotModule> = Object.create(null)
+// Assign browser's module cache and runtime modules to shared HMR state
+devModuleCache = Object.create(null)
 devContextPrototype.c = devModuleCache
+runtimeModules = new Set()
 
 // This file must not use `import` and `export` statements. Otherwise, it
 // becomes impossible to augment interfaces declared in `<reference>`d files
@@ -50,22 +51,6 @@ interface DevRuntimeBackend {
   restart: () => void
 }
 
-class UpdateApplyError extends Error {
-  name = 'UpdateApplyError'
-
-  dependencyChain: ModuleId[]
-
-  constructor(message: string, dependencyChain: ModuleId[]) {
-    super(message)
-    this.dependencyChain = dependencyChain
-  }
-}
-
-/**
- * Module IDs that are instantiated as part of the runtime of a chunk.
- */
-const runtimeModules: Set<ModuleId> = new Set()
-
 /**
  * Map from module ID to the chunks that contain this module.
  *
@@ -92,20 +77,6 @@ const chunkListChunksMap: Map<ChunkListPath, Set<ChunkPath>> = new Map()
  * Map from a chunk path to the chunk lists it belongs to.
  */
 const chunkChunkListsMap: Map<ChunkPath, Set<ChunkListPath>> = new Map()
-
-/**
- * Maps module IDs to persisted data between executions of their hot module
- * implementation (`hot.data`).
- */
-const moduleHotData: Map<ModuleId, HotData> = new Map()
-/**
- * Maps module instances to their hot module state.
- */
-const moduleHotState: Map<Module, HotState> = new Map()
-/**
- * Modules that call `module.hot.invalidate()` (while being updated).
- */
-const queuedInvalidatedModules: Set<ModuleId> = new Set()
 
 /**
  * Gets or instantiates a runtime module.
@@ -185,74 +156,36 @@ function instantiateModule(
   sourceType: SourceType,
   sourceData: SourceData
 ): Module {
-  // We are in development, this is always a string.
-  let id = moduleId as string
+  // Browser: creates base HotModule object (hot API added by shared code)
+  const createModuleObjectFn = (id: ModuleId) => {
+    return createModuleObject(id) as HotModule
+  }
 
-  const moduleFactory = moduleFactories.get(id)
-  if (typeof moduleFactory !== 'function') {
-    // This can happen if modules incorrectly handle HMR disposes/updates,
-    // e.g. when they keep a `setTimeout` around which still executes old code
-    // and contains e.g. a `require("something")` call.
-    throw new Error(
-      factoryNotAvailableMessage(id, sourceType, sourceData) +
-        ' It might have been deleted in an HMR update.'
+  // Browser: creates DevContext with refresh
+  const createContext = (
+    module: HotModule,
+    exports: Exports,
+    refresh: RefreshContext
+  ) => {
+    return new (DevContext as any as DevContextConstructor)(
+      module,
+      exports,
+      refresh
     )
   }
 
-  const hotData = moduleHotData.get(id)!
-  const { hot, hotState } = createModuleHot(id, hotData)
-
-  let parents: ModuleId[]
-  switch (sourceType) {
-    case SourceType.Runtime:
-      runtimeModules.add(id)
-      parents = []
-      break
-    case SourceType.Parent:
-      // No need to add this module as a child of the parent module here, this
-      // has already been taken care of in `getOrInstantiateModuleFromParent`.
-      parents = [sourceData as ModuleId]
-      break
-    case SourceType.Update:
-      parents = (sourceData as ModuleId[]) || []
-      break
-    default:
-      invariant(
-        sourceType,
-        (sourceType) => `Unknown source type: ${sourceType}`
-      )
-  }
-
-  const module: HotModule = createModuleObject(id) as HotModule
-  const exports = module.exports
-  module.parents = parents
-  module.children = []
-  module.hot = hot
-
-  devModuleCache[id] = module
-  moduleHotState.set(module, hotState)
-
-  // NOTE(alexkirsz) This can fail when the module encounters a runtime error.
-  try {
-    runModuleExecutionHooks(module, (refresh) => {
-      const context = new (DevContext as any as DevContextConstructor)(
-        module,
-        exports,
-        refresh
-      )
-      moduleFactory(context, module, exports)
-    })
-  } catch (error) {
-    module.error = error as any
-    throw error
-  }
-
-  if (module.namespaceObject && module.exports !== module.namespaceObject) {
-    // in case of a circular dependency: cjs1 -> esm2 -> cjs1
-    interopEsm(module.exports, module.namespaceObject)
-  }
-
-  return module
+  // Use shared instantiation logic (includes hot API setup)
+  return instantiateModuleShared(
+    moduleId,
+    sourceType,
+    sourceData,
+    moduleFactories,
+    devModuleCache,
+    runtimeModules,
+    createModuleObjectFn,
+    createContext,
+    runModuleExecutionHooks
+  )
 }
 
 const DUMMY_REFRESH_CONTEXT = {
@@ -349,91 +282,6 @@ function registerExportsAndSetupBoundaryForReactRefresh(
   }
 }
 
-function formatDependencyChain(dependencyChain: ModuleId[]): string {
-  return `Dependency chain: ${dependencyChain.join(' -> ')}`
-}
-
-function computeOutdatedModules(
-  added: Map<ModuleId, EcmascriptModuleEntry | undefined>,
-  modified: Map<ModuleId, EcmascriptModuleEntry>
-): {
-  outdatedModules: Set<ModuleId>
-  newModuleFactories: Map<ModuleId, ModuleFactory>
-} {
-  const newModuleFactories = new Map<ModuleId, ModuleFactory>()
-
-  for (const [moduleId, entry] of added) {
-    if (entry != null) {
-      newModuleFactories.set(moduleId, _eval(entry))
-    }
-  }
-
-  const outdatedModules = computedInvalidatedModules(modified.keys())
-
-  for (const [moduleId, entry] of modified) {
-    newModuleFactories.set(moduleId, _eval(entry))
-  }
-
-  return { outdatedModules, newModuleFactories }
-}
-
-function computedInvalidatedModules(
-  invalidated: Iterable<ModuleId>
-): Set<ModuleId> {
-  const outdatedModules = new Set<ModuleId>()
-
-  for (const moduleId of invalidated) {
-    const effect = getAffectedModuleEffects(moduleId)
-
-    switch (effect.type) {
-      case 'unaccepted':
-        throw new UpdateApplyError(
-          `cannot apply update: unaccepted module. ${formatDependencyChain(
-            effect.dependencyChain
-          )}.`,
-          effect.dependencyChain
-        )
-      case 'self-declined':
-        throw new UpdateApplyError(
-          `cannot apply update: self-declined module. ${formatDependencyChain(
-            effect.dependencyChain
-          )}.`,
-          effect.dependencyChain
-        )
-      case 'accepted':
-        for (const outdatedModuleId of effect.outdatedModules) {
-          outdatedModules.add(outdatedModuleId)
-        }
-        break
-      // TODO(alexkirsz) Dependencies: handle dependencies effects.
-      default:
-        invariant(effect, (effect) => `Unknown effect type: ${effect?.type}`)
-    }
-  }
-
-  return outdatedModules
-}
-
-function computeOutdatedSelfAcceptedModules(
-  outdatedModules: Iterable<ModuleId>
-): { moduleId: ModuleId; errorHandler: true | Function }[] {
-  const outdatedSelfAcceptedModules: {
-    moduleId: ModuleId
-    errorHandler: true | Function
-  }[] = []
-  for (const moduleId of outdatedModules) {
-    const module = devModuleCache[moduleId]
-    const hotState = moduleHotState.get(module)!
-    if (module && hotState.selfAccepted && !hotState.selfInvalidated) {
-      outdatedSelfAcceptedModules.push({
-        moduleId,
-        errorHandler: hotState.selfAccepted,
-      })
-    }
-  }
-  return outdatedSelfAcceptedModules
-}
-
 /**
  * Adds, deletes, and moves modules between chunks. This must happen before the
  * dispose phase as it needs to know which modules were removed from all chunks,
@@ -459,139 +307,6 @@ function updateChunksPhase(
   }
 
   return { disposedModules }
-}
-
-function disposePhase(
-  outdatedModules: Iterable<ModuleId>,
-  disposedModules: Iterable<ModuleId>
-): { outdatedModuleParents: Map<ModuleId, Array<ModuleId>> } {
-  for (const moduleId of outdatedModules) {
-    disposeModule(moduleId, 'replace')
-  }
-
-  for (const moduleId of disposedModules) {
-    disposeModule(moduleId, 'clear')
-  }
-
-  // Removing modules from the module cache is a separate step.
-  // We also want to keep track of previous parents of the outdated modules.
-  const outdatedModuleParents = new Map()
-  for (const moduleId of outdatedModules) {
-    const oldModule = devModuleCache[moduleId]
-    outdatedModuleParents.set(moduleId, oldModule?.parents)
-    delete devModuleCache[moduleId]
-  }
-
-  // TODO(alexkirsz) Dependencies: remove outdated dependency from module
-  // children.
-
-  return { outdatedModuleParents }
-}
-
-/**
- * Disposes of an instance of a module.
- *
- * Returns the persistent hot data that should be kept for the next module
- * instance.
- *
- * NOTE: mode = "replace" will not remove modules from the devModuleCache
- * This must be done in a separate step afterwards.
- * This is important because all modules need to be disposed to update the
- * parent/child relationships before they are actually removed from the devModuleCache.
- * If this was done in this method, the following disposeModule calls won't find
- * the module from the module id in the cache.
- */
-function disposeModule(moduleId: ModuleId, mode: 'clear' | 'replace') {
-  const module = devModuleCache[moduleId]
-  if (!module) {
-    return
-  }
-
-  const hotState = moduleHotState.get(module)!
-  const data = {}
-
-  // Run the `hot.dispose` handler, if any, passing in the persistent
-  // `hot.data` object.
-  for (const disposeHandler of hotState.disposeHandlers) {
-    disposeHandler(data)
-  }
-
-  // This used to warn in `getOrInstantiateModuleFromParent` when a disposed
-  // module is still importing other modules.
-  module.hot.active = false
-
-  moduleHotState.delete(module)
-
-  // TODO(alexkirsz) Dependencies: delete the module from outdated deps.
-
-  // Remove the disposed module from its children's parent list.
-  // It will be added back once the module re-instantiates and imports its
-  // children again.
-  for (const childId of module.children) {
-    const child = devModuleCache[childId]
-    if (!child) {
-      continue
-    }
-
-    const idx = child.parents.indexOf(module.id)
-    if (idx >= 0) {
-      child.parents.splice(idx, 1)
-    }
-  }
-
-  switch (mode) {
-    case 'clear':
-      delete devModuleCache[module.id]
-      moduleHotData.delete(module.id)
-      break
-    case 'replace':
-      moduleHotData.set(module.id, data)
-      break
-    default:
-      invariant(mode, (mode) => `invalid mode: ${mode}`)
-  }
-}
-
-function applyPhase(
-  outdatedSelfAcceptedModules: {
-    moduleId: ModuleId
-    errorHandler: true | Function
-  }[],
-  newModuleFactories: Map<ModuleId, ModuleFactory>,
-  outdatedModuleParents: Map<ModuleId, Array<ModuleId>>,
-  reportError: (err: any) => void
-) {
-  // Update module factories.
-  for (const [moduleId, factory] of newModuleFactories.entries()) {
-    applyModuleFactoryName(factory)
-    moduleFactories.set(moduleId, factory)
-  }
-
-  // TODO(alexkirsz) Run new runtime entries here.
-
-  // TODO(alexkirsz) Dependencies: call accept handlers for outdated deps.
-
-  // Re-instantiate all outdated self-accepted modules.
-  for (const { moduleId, errorHandler } of outdatedSelfAcceptedModules) {
-    try {
-      instantiateModule(
-        moduleId,
-        SourceType.Update,
-        outdatedModuleParents.get(moduleId)
-      )
-    } catch (err) {
-      if (typeof errorHandler === 'function') {
-        try {
-          errorHandler(err, { moduleId, module: devModuleCache[moduleId] })
-        } catch (err2) {
-          reportError(err2)
-          reportError(err)
-        }
-      } else {
-        reportError(err)
-      }
-    }
-  }
 }
 
 function applyUpdate(update: PartialUpdate) {
@@ -651,257 +366,30 @@ function applyChunkListUpdate(update: ChunkListUpdate) {
 }
 
 function applyEcmascriptMergedUpdate(update: EcmascriptMergedUpdate) {
+  // Browser-specific chunk management phase
   const { entries = {}, chunks = {} } = update
   const { added, modified, chunksAdded, chunksDeleted } = computeChangedModules(
     entries,
-    chunks
-  )
-  const { outdatedModules, newModuleFactories } = computeOutdatedModules(
-    added,
-    modified
+    chunks,
+    chunkModulesMap
   )
   const { disposedModules } = updateChunksPhase(chunksAdded, chunksDeleted)
 
-  applyInternal(outdatedModules, disposedModules, newModuleFactories)
-}
-
-function applyInvalidatedModules(outdatedModules: Set<ModuleId>) {
-  if (queuedInvalidatedModules.size > 0) {
-    computedInvalidatedModules(queuedInvalidatedModules).forEach((moduleId) => {
-      outdatedModules.add(moduleId)
-    })
-
-    queuedInvalidatedModules.clear()
-  }
-
-  return outdatedModules
-}
-
-function applyInternal(
-  outdatedModules: Set<ModuleId>,
-  disposedModules: Iterable<ModuleId>,
-  newModuleFactories: Map<ModuleId, ModuleFactory>
-) {
-  outdatedModules = applyInvalidatedModules(outdatedModules)
-
-  const outdatedSelfAcceptedModules =
-    computeOutdatedSelfAcceptedModules(outdatedModules)
-
-  const { outdatedModuleParents } = disposePhase(
-    outdatedModules,
-    disposedModules
+  // Use shared HMR update implementation
+  applyEcmascriptMergedUpdateShared(
+    added,
+    modified,
+    disposedModules,
+    _eval, // browser's eval with source maps
+    instantiateModule, // now wraps shared logic
+    applyModuleFactoryName,
+    moduleFactories,
+    devModuleCache,
+    runtimeModules
   )
-
-  // we want to continue on error and only throw the error after we tried applying all updates
-  let error: any
-
-  function reportError(err: any) {
-    if (!error) error = err
-  }
-
-  applyPhase(
-    outdatedSelfAcceptedModules,
-    newModuleFactories,
-    outdatedModuleParents,
-    reportError
-  )
-
-  if (error) {
-    throw error
-  }
-
-  if (queuedInvalidatedModules.size > 0) {
-    applyInternal(new Set(), [], new Map())
-  }
 }
 
-function computeChangedModules(
-  entries: Record<ModuleId, EcmascriptModuleEntry>,
-  updates: Record<ChunkPath, EcmascriptMergedChunkUpdate>
-): {
-  added: Map<ModuleId, EcmascriptModuleEntry | undefined>
-  modified: Map<ModuleId, EcmascriptModuleEntry>
-  deleted: Set<ModuleId>
-  chunksAdded: Map<ChunkPath, Set<ModuleId>>
-  chunksDeleted: Map<ChunkPath, Set<ModuleId>>
-} {
-  const chunksAdded = new Map()
-  const chunksDeleted = new Map()
-  const added: Map<ModuleId, EcmascriptModuleEntry> = new Map()
-  const modified = new Map()
-  const deleted: Set<ModuleId> = new Set()
-
-  for (const [chunkPath, mergedChunkUpdate] of Object.entries(updates) as Array<
-    [ChunkPath, EcmascriptMergedChunkUpdate]
-  >) {
-    switch (mergedChunkUpdate.type) {
-      case 'added': {
-        const updateAdded = new Set(mergedChunkUpdate.modules)
-        for (const moduleId of updateAdded) {
-          added.set(moduleId, entries[moduleId])
-        }
-        chunksAdded.set(chunkPath, updateAdded)
-        break
-      }
-      case 'deleted': {
-        // We could also use `mergedChunkUpdate.modules` here.
-        const updateDeleted = new Set(chunkModulesMap.get(chunkPath))
-        for (const moduleId of updateDeleted) {
-          deleted.add(moduleId)
-        }
-        chunksDeleted.set(chunkPath, updateDeleted)
-        break
-      }
-      case 'partial': {
-        const updateAdded = new Set(mergedChunkUpdate.added)
-        const updateDeleted = new Set(mergedChunkUpdate.deleted)
-        for (const moduleId of updateAdded) {
-          added.set(moduleId, entries[moduleId])
-        }
-        for (const moduleId of updateDeleted) {
-          deleted.add(moduleId)
-        }
-        chunksAdded.set(chunkPath, updateAdded)
-        chunksDeleted.set(chunkPath, updateDeleted)
-        break
-      }
-      default:
-        invariant(
-          mergedChunkUpdate,
-          (mergedChunkUpdate) =>
-            `Unknown merged chunk update type: ${mergedChunkUpdate.type}`
-        )
-    }
-  }
-
-  // If a module was added from one chunk and deleted from another in the same update,
-  // consider it to be modified, as it means the module was moved from one chunk to another
-  // AND has new code in a single update.
-  for (const moduleId of added.keys()) {
-    if (deleted.has(moduleId)) {
-      added.delete(moduleId)
-      deleted.delete(moduleId)
-    }
-  }
-
-  for (const [moduleId, entry] of Object.entries(entries)) {
-    // Modules that haven't been added to any chunk but have new code are considered
-    // to be modified.
-    // This needs to be under the previous loop, as we need it to get rid of modules
-    // that were added and deleted in the same update.
-    if (!added.has(moduleId)) {
-      modified.set(moduleId, entry)
-    }
-  }
-
-  return { added, deleted, modified, chunksAdded, chunksDeleted }
-}
-
-type ModuleEffect =
-  | {
-      type: 'unaccepted'
-      dependencyChain: ModuleId[]
-    }
-  | {
-      type: 'self-declined'
-      dependencyChain: ModuleId[]
-      moduleId: ModuleId
-    }
-  | {
-      type: 'accepted'
-      moduleId: ModuleId
-      outdatedModules: Set<ModuleId>
-    }
-
-function getAffectedModuleEffects(moduleId: ModuleId): ModuleEffect {
-  const outdatedModules: Set<ModuleId> = new Set()
-
-  type QueueItem = { moduleId?: ModuleId; dependencyChain: ModuleId[] }
-
-  const queue: QueueItem[] = [
-    {
-      moduleId,
-      dependencyChain: [],
-    },
-  ]
-
-  let nextItem
-  while ((nextItem = queue.shift())) {
-    const { moduleId, dependencyChain } = nextItem
-
-    if (moduleId != null) {
-      if (outdatedModules.has(moduleId)) {
-        // Avoid infinite loops caused by cycles between modules in the dependency chain.
-        continue
-      }
-
-      outdatedModules.add(moduleId)
-    }
-
-    // We've arrived at the runtime of the chunk, which means that nothing
-    // else above can accept this update.
-    if (moduleId === undefined) {
-      return {
-        type: 'unaccepted',
-        dependencyChain,
-      }
-    }
-
-    const module = devModuleCache[moduleId]
-    const hotState = moduleHotState.get(module)!
-
-    if (
-      // The module is not in the cache. Since this is a "modified" update,
-      // it means that the module was never instantiated before.
-      !module || // The module accepted itself without invalidating globalThis.
-      // TODO is that right?
-      (hotState.selfAccepted && !hotState.selfInvalidated)
-    ) {
-      continue
-    }
-
-    if (hotState.selfDeclined) {
-      return {
-        type: 'self-declined',
-        dependencyChain,
-        moduleId,
-      }
-    }
-
-    if (runtimeModules.has(moduleId)) {
-      queue.push({
-        moduleId: undefined,
-        dependencyChain: [...dependencyChain, moduleId],
-      })
-      continue
-    }
-
-    for (const parentId of module.parents) {
-      const parent = devModuleCache[parentId]
-
-      if (!parent) {
-        // TODO(alexkirsz) Is this even possible?
-        continue
-      }
-
-      // TODO(alexkirsz) Dependencies: check accepted and declined
-      // dependencies here.
-
-      queue.push({
-        moduleId: parentId,
-        dependencyChain: [...dependencyChain, moduleId],
-      })
-    }
-  }
-
-  return {
-    type: 'accepted',
-    moduleId,
-    outdatedModules,
-  }
-}
-
-function handleApply(chunkListPath: ChunkListPath, update: ServerMessage) {
+function handleApply(chunkListPath: ChunkListPath, update: HMRPayload) {
   switch (update.type) {
     case 'partial': {
       // This indicates that the update is can be applied to the current state of the application.
@@ -930,86 +418,6 @@ function handleApply(chunkListPath: ChunkListPath, update: ServerMessage) {
     default:
       throw new Error(`Unknown update type: ${update.type}`)
   }
-}
-
-function createModuleHot(
-  moduleId: ModuleId,
-  hotData: HotData
-): { hot: Hot; hotState: HotState } {
-  const hotState: HotState = {
-    selfAccepted: false,
-    selfDeclined: false,
-    selfInvalidated: false,
-    disposeHandlers: [],
-  }
-
-  const hot: Hot = {
-    // TODO(alexkirsz) This is not defined in the HMR API. It was used to
-    // decide whether to warn whenever an HMR-disposed module required other
-    // modules. We might want to remove it.
-    active: true,
-
-    data: hotData ?? {},
-
-    // TODO(alexkirsz) Support full (dep, callback, errorHandler) form.
-    accept: (
-      modules?: string | string[] | AcceptErrorHandler,
-      _callback?: AcceptCallback,
-      _errorHandler?: AcceptErrorHandler
-    ) => {
-      if (modules === undefined) {
-        hotState.selfAccepted = true
-      } else if (typeof modules === 'function') {
-        hotState.selfAccepted = modules
-      } else {
-        throw new Error('unsupported `accept` signature')
-      }
-    },
-
-    decline: (dep) => {
-      if (dep === undefined) {
-        hotState.selfDeclined = true
-      } else {
-        throw new Error('unsupported `decline` signature')
-      }
-    },
-
-    dispose: (callback) => {
-      hotState.disposeHandlers.push(callback)
-    },
-
-    addDisposeHandler: (callback) => {
-      hotState.disposeHandlers.push(callback)
-    },
-
-    removeDisposeHandler: (callback) => {
-      const idx = hotState.disposeHandlers.indexOf(callback)
-      if (idx >= 0) {
-        hotState.disposeHandlers.splice(idx, 1)
-      }
-    },
-
-    invalidate: () => {
-      hotState.selfInvalidated = true
-      queuedInvalidatedModules.add(moduleId)
-    },
-
-    // NOTE(alexkirsz) This is part of the management API, which we don't
-    // implement, but the Next.js React Refresh runtime uses this to decide
-    // whether to schedule an update.
-    status: () => 'idle',
-
-    // NOTE(alexkirsz) Since we always return "idle" for now, these are no-ops.
-    addStatusHandler: (_handler) => {},
-    removeStatusHandler: (_handler) => {},
-
-    // NOTE(jridgewell) Check returns the list of updated modules, but we don't
-    // want the webpack code paths to ever update (the turbopack paths handle
-    // this already).
-    check: () => Promise.resolve(null),
-  }
-
-  return { hot, hotState }
 }
 
 /**
